training-model/requirements.txt
ultralytics
roboflow
pyyaml
